requirements.txt
fastapi==0.115.8
google-auth==2.38.0
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
python-multipart==0.0.20
requests==2.32.3
uvicorn==0.34.0
pydantic[email]